export config setup as a module
since the update where config reads user configuration from the
database and other modules immediately depend on that configuration, it
is important to ensure that configurations have loaded before trying to
require those other modules.

the setup method will be used as a blocking call to load configurations
before starting any of the server methods

it has been determined that a blocking call in setup/configuration is
okay, since it will not affect general runtime performance
the setup method is not included in config, as config should only export
a javascript object containing the configuration information
